Итоговая версия

+++ ConferenceApp.Core/Repositories/RequestRepository.cs
@@ -17,7 +17,7 @@ namespace ConferenceApp.Core.Repositories
 
         public RequestRepository
         (
-            IReportRepository reportRepository, 
+            IReportRepository reportRepository,
             IUserRepository userRepository,
             MainDb db
         )
@@ -26,8 +26,8 @@ namespace ConferenceApp.Core.Repositories
             _userRepository = userRepository;
             _db = db;
         }
-        
-        
+
+
         /// <summary>
         /// Добавить заявку.
         /// </summary>
@@ -35,58 +35,95 @@ namespace ConferenceApp.Core.Repositories
         {
             // 1. Добавить пользователя.
             var user = model.User;
-            var userId = _userRepository.InsertWithId( user );
+            var userId = _userRepository.InsertWithId(user);
 
             // 2. Добавить заявку.
-            var request = new Request { Id = Guid.NewGuid(), OwnerId = userId };
-            _db.Insert( request );
+            var request = new Request {Id = Guid.NewGuid(), OwnerId = userId};
+            _db.Insert(request);
 
             // 3. Добавить доклады.
             var reports = model.Reports;
-            _reportRepository.InsertRange( reports );
+            _reportRepository.InsertRange(reports);
         }
 
 
         /// <summary>
         /// Удалить заявку.
         /// </summary>
-        public void Delete(Guid userId)
+        public void Delete( Guid userId )
         {
             var request = _db.Requests.FirstOrDefault(x => x.OwnerId == userId);
+            if( request == null )
+            {
+                return;
+            }
 
             // 1. Удалить доклады.
-            _reportRepository.DeleteRange( request.Id );
+            _reportRepository.DeleteRange(request.Id);
 
             // 2. Удалить заявку.
-            _db.Requests.Delete( x => x.Id == userId );
+            _db.Requests.Delete(x => x.Id == userId);
 
             // 3. Удалить пользователя.
-            _userRepository.Delete( request.OwnerId );
+            _userRepository.Delete(request.OwnerId);
         }
 
 
         /// <summary>
         /// Обновить информацию по заявке.
         /// </summary>
-        public void Update(RequestModel request)
+        public void Update( RequestModel request )
         {
             throw new NotImplementedException();
+        }
+
+
+        public void ChangeStatus( Guid requestId, RequestStatus status )
+        {
+            _db.Requests
+                .Where(x => x.Id == requestId)
+                .Set(x => x.Status, status)
+                .Update();
         }
 
 
         /// <summary>
         /// Получить заявку по id владельца.
         /// </summary>
-        public RequestModel Get(Guid userId)
+        public RequestModel Get( Guid requestId )
         {
-            throw new NotImplementedException();
+            var request = GetDto(requestId);
+            if( request == null )
+            {
+                return null;
+            }
+
+            var model = new RequestModel
+            {
+                User = _userRepository
+                    .Get(x => x.Id == request.OwnerId)
+                    .FirstOrDefault(),
+                Reports = _reportRepository
+                    .GetReportsByRequest(request.Id)
+                    .ToList()
+            };
+            return model;
+        }
+
+
+        /// <summary>
+        /// Вернуть dto.
+        /// </summary>
+        public Request GetDto( Guid requestId )
+        {
+            return _db.Requests.FirstOrDefault(x => x.Id == requestId);
         }
 
 
         /// <summary>
         /// Получить заявки по критерию.
         /// </summary>
-        public IEnumerable<RequestModel> Get(Func<RequestModel, bool> filter)
+        public IEnumerable<RequestModel> Get( Func<RequestModel, bool> filter )
         {
             throw new NotImplementedException();
         }
@@ -97,7 +134,18 @@ namespace ConferenceApp.Core.Repositories
         /// </summary>
         public IEnumerable<RequestModel> GetAll()
         {
-            throw new NotImplementedException();
+            var requests = _db.Requests.ToList();
+            var model = requests.Select(request => new RequestModel
+                {
+                    User = _userRepository
+                        .Get(x => x.Id == request.OwnerId)
+                        .FirstOrDefault(),
+                    Reports = _reportRepository
+                        .GetReportsByRequest(request.Id)
+                        .ToList()
+                })
+                .ToList();
+            return model;
         }
     }
 }
